// src/generated/TransactionBuilder.ts
export interface TransactionBuilderDef {
    applyForSR: (address: string | null, url: string | null, callParams: CallParams<'address' | 'url'>) => string | Promise<string>;
    clearABI: (address: string, ownerAddress: string, callParams: CallParams<'address' | 'ownerAddress'>) => string | Promise<string>;
    createTRXExchange: (tokenName: string, tokenBalance: number, trxBalance: number, ownerAddress: string, callParams: CallParams<'tokenName' | 'tokenBalance' | 'trxBalance' | 'ownerAddress'>) => string | Promise<string>;
    createTokenExchange: (firstTokenName: string, firstTokenBalance: number, secondTokenName: string, secondTokenBalance: number, ownerAddress: string, callParams: CallParams<'firstTokenName' | 'firstTokenBalance' | 'secondTokenName' | 'secondTokenBalance' | 'ownerAddress'>) => string | Promise<string>;
    deleteProposal: (proposalID: number, issuerAddress: string, callParams: CallParams<'proposalID' | 'issuerAddress'>) => string | Promise<string>;
    freezeBalance: (receiverAddress: string, amount: number | null, duration: number | null, resource: string | null, address: string | null, callParams: CallParams<'receiverAddress' | 'amount' | 'duration' | 'resource' | 'address'>) => string | Promise<string>;
    injectExchangeTokens: (exchangeID: number, tokenName: string, tokenAmount: number, ownerAddress: string, callParams: CallParams<'exchangeID' | 'tokenName' | 'tokenAmount' | 'ownerAddress'>) => string | Promise<string>;
    purchaseAsset: (issuerAddresso: string, tokenID: string, amount: number, buyer: string | null, callParams: CallParams<'issuerAddresso' | 'tokenID' | 'amount' | 'buyer'>) => string | Promise<string>;
    purchaseToken: (issuerAddresso: string, tokenID: string, amount: number, buyer: string | null, callParams: CallParams<'issuerAddresso' | 'tokenID' | 'amount' | 'buyer'>) => string | Promise<string>;
    sendAsset: (to: string, amount: number, tokenID: string, fromAddr: string | null, callParams: CallParams<'to' | 'amount' | 'tokenID' | 'fromAddr'>) => string | Promise<string>;
    sendToken: (to: string, amount: number, tokenID: string, fromAddr: string | null, callParams: CallParams<'to' | 'amount' | 'tokenID' | 'fromAddr'>) => string | Promise<string>;
    sendTrx: (to: string, amount: number, fromAddr: string | null, callParams: CallParams<'to' | 'amount' | 'fromAddr'>) => string | Promise<string>;
    setAccountId: (accountName: string, accountId: string, callParams: CallParams<'accountName' | 'accountId'>) => string | Promise<string>;
    tradeExchangeTokens: (exchangeID: number, tokenName: string, tokenAmountSold: number, tokenAmountExpected: number, ownerAddress: string, callParams: CallParams<'exchangeID' | 'tokenName' | 'tokenAmountSold' | 'tokenAmountExpected' | 'ownerAddress'>) => string | Promise<string>;
    unfreezeBalance: (receiverAddress: string, resource: string | null, address: string | null, callParams: CallParams<'receiverAddress' | 'resource' | 'address'>) => string | Promise<string>;
    updateAccount: (accountName: string, address: string | null, callParams: CallParams<'accountName' | 'address'>) => string | Promise<string>;
    updateBrokerage: (brokerage: number, ownerAddress: string, callParams: CallParams<'brokerage' | 'ownerAddress'>) => string | Promise<string>;
    updateEnergyLimit: (contractAddress: string, originEnergyLimit: number, ownerAddress: string, callParams: CallParams<'contractAddress' | 'originEnergyLimit' | 'ownerAddress'>) => string | Promise<string>;
    updateSetting: (contractAddress: string, userFeePercentage: number, ownerAddress: string, callParams: CallParams<'contractAddress' | 'userFeePercentage' | 'ownerAddress'>) => string | Promise<string>;
    voteProposal: (proposalID: number, isApproval: boolean, voterAddress: string, callParams: CallParams<'proposalID' | 'isApproval' | 'voterAddress'>) => string | Promise<string>;
    withdrawBlockRewards: (address: string, callParams: CallParams<'address'>) => string | Promise<string>;
    withdrawExchangeTokens: (exchangeID: number, tokenName: string, tokenAmount: number, ownerAddress: string, callParams: CallParams<'exchangeID' | 'tokenName' | 'tokenAmount' | 'ownerAddress'>) => string | Promise<string>;
}